+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-StarWars-client/src/main/java/com/graphql_java_generator/samples/simple/client/Main.java
@@ -1,22 +1,33 @@
 package com.graphql_java_generator.samples.simple.client;
 
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.security.cert.X509Certificate;
+import java.util.Collections;
 
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLSession;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
+import javax.net.ssl.SSLException;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.client.reactive.ReactorClientHttpConnector;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.WebClient.Builder;
 
 import com.generated.graphql.Episode;
+import com.generated.graphql.QueryTypeExecutor;
+import com.graphql_java_generator.client.GraphQLConfiguration;
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
 import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
 import com.graphql_java_generator.samples.simple.client.graphql.PartialDirectRequests;
 import com.graphql_java_generator.samples.simple.client.graphql.PartialPreparedRequests;
+
+import io.netty.handler.ssl.SslContext;
+import io.netty.handler.ssl.SslContextBuilder;
+import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
+import reactor.netty.http.client.HttpClient;
 
 /**
  * The main class, which executes the same queries, built by three different methods. See {@link PartialDirectRequests},
@@ -24,27 +35,41 @@ import com.graphql_java_generator.samples.simple.client.graphql.PartialPreparedR
  * 
  * @author etienne-sf
  */
-public class Main {
+@SpringBootApplication(scanBasePackageClasses = { Main.class, GraphQLConfiguration.class, QueryTypeExecutor.class })
+public class Main implements CommandLineRunner {
 
-	public static String graphqlEndpoint = "https://localhost:8443/starwars/graphql";
+	@Value("${graphql.endpoint.url}")
+	private String graphqlEndpoint;
+	@Value("${graphql.endpoint.subscriptionUrl}")
+	private String graphqlSubscriptionEndpoint;
+
+	@Autowired
+	PartialDirectRequests partialDirectRequests;
+	@Autowired
+	PartialPreparedRequests partialPreparedRequests;
 
 	public static void main(String[] args) throws Exception {
-		new Main().execAll();
+		SpringApplication.run(Main.class, args);
 	}
 
-	void execAll() throws Exception {
+	/**
+	 * This method is started by Spring, once the Spring context has been loaded. This is run, as this class implements
+	 * {@link CommandLineRunner}
+	 */
+	@Override
+	public void run(String... args) throws Exception {
 
-		// Execution of three way to user the GraphQL client, to call the GraphQL server
+		// Execution of two ways that use the GraphQL client, to call the GraphQL server
 
 		System.out.println("============================================================================");
 		System.out.println("======= SIMPLEST WAY: DIRECT QUERIES =======================================");
 		System.out.println("============================================================================");
-		execOne(new PartialDirectRequests(graphqlEndpoint, getNoCheckSslContext(), new NoOpHostnameVerifier()));
+		execOne(partialDirectRequests);
 
 		System.out.println("============================================================================");
 		System.out.println("======= MOST SECURE WAY: PREPARED QUERIES ==================================");
 		System.out.println("============================================================================");
-		execOne(new PartialPreparedRequests(graphqlEndpoint, getNoCheckSslContext(), new NoOpHostnameVerifier()));
+		execOne(partialPreparedRequests);
 
 		System.out.println("");
 		System.out.println("");
@@ -104,6 +129,40 @@ public class Main {
 		}
 	}
 
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+	////////////////// Below is the configuration, based on Spring beans
+	//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+	@Bean
+	String graphqlEndpoint() {
+		return graphqlEndpoint;
+	}
+
+	@Bean
+	String graphqlSubscriptionEndpoint() {
+		return graphqlSubscriptionEndpoint;
+	}
+
+	@Bean
+	SslContext insecureSslContext() throws SSLException {
+		return SslContextBuilder.forClient().trustManager(InsecureTrustManagerFactory.INSTANCE).build();
+	}
+
+	@Bean
+	WebClient webClient(String graphqlEndpoint, SslContext nettySslContext) {
+		Builder ret = WebClient.builder()//
+				.baseUrl(graphqlEndpoint)//
+				// .defaultCookie("cookieKey", "cookieValue")//
+				.defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+				.defaultUriVariables(Collections.singletonMap("url", graphqlEndpoint));
+
+		if (nettySslContext != null) {
+			HttpClient httpConnector = HttpClient.create().secure(t -> t.sslContext(nettySslContext));
+			ret.clientConnector(new ReactorClientHttpConnector(httpConnector));
+		}
+		return ret.build();
+	}
+
 	/**
 	 * Returns a Dummy SSLContext, that won't check the server certificate. With a real check, an exception is thrown,
 	 * as it is self signed.<BR/>
@@ -113,48 +172,49 @@ public class Main {
 	 * @throws NoSuchAlgorithmException
 	 * @throws KeyManagementException
 	 */
-	public SSLContext getNoCheckSslContext() throws NoSuchAlgorithmException, KeyManagementException {
-		SSLContext sslContext = SSLContext.getInstance("TLSv1");
-
-		// Very, very bad. Don't do that in production !
-		KeyManager[] keyManagers = null;
-		TrustManager[] trustManager = { new NoOpTrustManager() };
-		SecureRandom secureRandom = new SecureRandom();
-
-		sslContext.init(keyManagers, trustManager, secureRandom);
-
-		return sslContext;
-	}
-
-	/**
-	 * A dummy checker. DON'T USE IT IN PRODUCTION. But we can't buy a real certificat for this project.
-	 * 
-	 * @return
-	 */
-	public HostnameVerifier getHostnameVerifier() {
-		// Very, very bad. Don't do that in production !
-		return new NoOpHostnameVerifier();
-	}
-
-	public class NoOpTrustManager implements X509TrustManager {
-		@Override
-		public void checkClientTrusted(X509Certificate[] x509Certificates, String s) {
-		}
-
-		@Override
-		public void checkServerTrusted(X509Certificate[] x509Certificates, String s) {
-		}
-
-		@Override
-		public X509Certificate[] getAcceptedIssuers() {
-			return new X509Certificate[0];
-		}
-	}
-
-	public class NoOpHostnameVerifier implements HostnameVerifier {
-		@Override
-		public boolean verify(String s, SSLSession sslSession) {
-			return true;
-		}
-	}
+	// @Bean
+	// public SSLContext sslContext() throws NoSuchAlgorithmException, KeyManagementException {
+	// SSLContext sslContext = SSLContext.getInstance("TLSv1");
+	//
+	// // Very, very bad. Don't do that in production !
+	// KeyManager[] keyManagers = null;
+	// TrustManager[] trustManager = { new X509TrustManager() {
+	// @Override
+	// public void checkClientTrusted(X509Certificate[] x509Certificates, String s) {
+	// }
+	//
+	// @Override
+	// public void checkServerTrusted(X509Certificate[] x509Certificates, String s) {
+	// }
+	//
+	// @Override
+	// public X509Certificate[] getAcceptedIssuers() {
+	// return new X509Certificate[0];
+	// }
+	// } };
+	// SecureRandom secureRandom = new SecureRandom();
+	//
+	// sslContext.init(keyManagers, trustManager, secureRandom);
+	//
+	// return sslContext;
+	// }
+	//
+	// /**
+	// * Creates a netty {@link SslContext} from a JDK {@link SSLContext}. It is called only if no netty
+	// * {@link SslContext} has already been defined.
+	// *
+	// * @return a default netty {@link SslContext}, if no JDK {@link SSLContext} Spring bean exists. And a netty
+	// * {@link SslContext} based on the existing JDK {@link SSLContext} Spring bean otherwise.
+	// * @throws SSLException
+	// */
+	// @ConditionalOnMissingBean
+	// @Bean
+	// SslContext nettySslContext(SSLContext sslContext) throws SSLException {
+	// if (sslContext == null) {
+	// return SslContextBuilder.forClient().build();
+	// } else {
+	// // return new JdkSslContext(sslContext, true, null, null, null, ClientAuth.NONE, null, true);
+	// return new JdkSslContext(sslContext, true, ClientAuth.NONE);
+	// }
+	// }
 }