+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-StarWars-client/src/main/java/com/graphql_java_generator/samples/simple/client/graphql/PartialDirectRequests.java
@@ -1,14 +1,14 @@
 package com.graphql_java_generator.samples.simple.client.graphql;
 
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.SSLContext;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import com.generated.graphql.Character;
 import com.generated.graphql.Droid;
 import com.generated.graphql.Episode;
 import com.generated.graphql.Human;
-import com.generated.graphql.MutationType;
-import com.generated.graphql.QueryType;
+import com.generated.graphql.MutationTypeExecutor;
+import com.generated.graphql.QueryTypeExecutor;
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
 import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
 import com.graphql_java_generator.samples.simple.client.Queries;
@@ -18,92 +18,81 @@ import com.graphql_java_generator.samples.simple.client.Queries;
  * 
  * @author etienne-sf
  */
+@Component
 public class PartialDirectRequests implements Queries {
 
-	final QueryType queryType;
-	final MutationType mutationType;
+	@Autowired
+	QueryTypeExecutor queryTypeExecutor;
 
-	/**
-	 * This constructor expects the URI of the GraphQL server. This constructor works only for http servers, not for
-	 * https ones.<BR/>
-	 * For example: https://my.server.com/graphql
-	 * 
-	 * @param graphqlEndpoint
-	 *            the https URI for the GraphQL endpoint
-	 * @param sslContext
-	 * @param hostnameVerifier
-	 */
-	public PartialDirectRequests(String graphqlEndpoint, SSLContext sslContext, HostnameVerifier hostnameVerifier) {
-		queryType = new QueryType(graphqlEndpoint, sslContext, hostnameVerifier);
-		mutationType = new MutationType(graphqlEndpoint, sslContext, hostnameVerifier);
-	}
+	@Autowired
+	MutationTypeExecutor mutationTypeExecutor;
 
 	@Override
 	public Character heroFull() throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.hero("", null);
+		return queryTypeExecutor.hero("", null);
 	}
 
 	@Override
 	public Character heroPartial(Episode episode)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.hero("{appearsIn name}", episode);
+		return queryTypeExecutor.hero("{appearsIn name}", episode);
 	}
 
 	@Override
 	public Character heroFriendsFriendsFriends(Episode episode)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.hero("{id appearsIn friends {name friends {friends{id name appearsIn}}}}", episode);
+		return queryTypeExecutor.hero("{id appearsIn friends {name friends {friends{id name appearsIn}}}}", episode);
 	}
 
 	@Override
 	public Human humanFull(String id) throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.human((String) null, id);
+		return queryTypeExecutor.human((String) null, id);
 	}
 
 	@Override
 	public Human humanPartial(String id) throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.human("{appearsIn homePlanet name}", id);
+		return queryTypeExecutor.human("{appearsIn homePlanet name}", id);
 	}
 
 	@Override
 	public Human humanFriendsFriendsFriends(String id)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.human("{id appearsIn name friends {name friends {friends{id name appearsIn}}}}", id);
+		return queryTypeExecutor.human("{id appearsIn name friends {name friends {friends{id name appearsIn}}}}", id);
 	}
 
 	@Override
 	public Droid droidFull(String id) throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.droid("", id);
+		return queryTypeExecutor.droid("", id);
 	}
 
 	@Override
 	public Droid droidPartial(String id) throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.droid("{appearsIn primaryFunction name}", id);
+		return queryTypeExecutor.droid("{appearsIn primaryFunction name}", id);
 	}
 
 	@Override
 	public Droid droidFriendsFriendsFriends(String id)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType
+		return queryTypeExecutor
 				.droid("{id appearsIn name friends {name friends {friends{id name appearsIn}}} primaryFunction }", id);
 	}
 
 	@Override
 	public Droid droidDoesNotExist() throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return queryType.droid("{id appearsIn friends {name friends {friends{id name appearsIn}}} primaryFunction }",
+		return queryTypeExecutor.droid("{id appearsIn friends {name friends {friends{id name appearsIn}}} primaryFunction }",
 				"00000000-0000-0000-0000-000000001111");
 	}
 
 	@Override
 	public Human createHuman(String name, String homePlanet)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return mutationType.createHuman("{id name appearsIn homePlanet friends {id name}}", name, homePlanet);
+		return mutationTypeExecutor.createHuman("{id name appearsIn homePlanet friends {id name}}", name, homePlanet);
 	}
 
 	@Override
 	public Character addFriend(String idCharacter, String idNewFriend)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return mutationType.addFriend("{id name appearsIn friends {id name}}", idCharacter, idNewFriend);
+		return mutationTypeExecutor.addFriend("{id name appearsIn friends {id name}}", idCharacter, idNewFriend);
 	}
 
 }