+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-StarWars-client/src/main/java/com/graphql_java_generator/samples/simple/client/graphql/PartialPreparedRequests.java
@@ -3,16 +3,18 @@
  */
 package com.graphql_java_generator.samples.simple.client.graphql;
 
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.SSLContext;
+import javax.annotation.PostConstruct;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import com.generated.graphql.Character;
 import com.generated.graphql.Droid;
 import com.generated.graphql.Episode;
 import com.generated.graphql.GraphQLRequest;
 import com.generated.graphql.Human;
-import com.generated.graphql.MutationType;
-import com.generated.graphql.QueryType;
+import com.generated.graphql.MutationTypeExecutor;
+import com.generated.graphql.QueryTypeExecutor;
 import com.graphql_java_generator.client.request.ObjectResponse;
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
 import com.graphql_java_generator.exception.GraphQLRequestPreparationException;
@@ -31,10 +33,13 @@ import com.graphql_java_generator.samples.simple.client.Queries;
  * 
  * @author etienne-sf
  */
+@Component
 public class PartialPreparedRequests implements Queries {
 
-	final QueryType queryType;
-	final MutationType mutationType;
+	@Autowired
+	QueryTypeExecutor queryTypeExecutor;
+	@Autowired
+	MutationTypeExecutor mutationTypeExecutor;
 
 	GraphQLRequest heroFullRequest;
 	GraphQLRequest heroPartialRequest;
@@ -52,43 +57,33 @@ public class PartialPreparedRequests implements Queries {
 	GraphQLRequest addFriendRequest;
 
 	/**
-	 * This constructor expects the URI of the GraphQL server. This constructor works only for http servers, not for
-	 * https ones.<BR/>
-	 * For example: https://my.server.com/graphql
-	 * 
-	 * @param graphqlEndpoint
-	 *            the https URI for the GraphQL endpoint
-	 * @param sslContext
-	 * @param hostnameVerifier
-	 * @throws GraphQLRequestPreparationException
+	 * This method is executed, once all dependencies have been loaded. It prepares all the GraphQL requests
 	 */
-	public PartialPreparedRequests(String graphqlEndpoint, SSLContext sslContext, HostnameVerifier hostnameVerifier)
-			throws GraphQLRequestPreparationException {
-		queryType = new QueryType(graphqlEndpoint, sslContext, hostnameVerifier);
-		mutationType = new MutationType(graphqlEndpoint, sslContext, hostnameVerifier);
+	@PostConstruct
+	public void init() throws GraphQLRequestPreparationException {
 
 		// The easiest way: don't precise which fields you want, and all known scalar fields are queried
-		heroFullRequest = queryType.getHeroGraphQLRequest(null);
+		heroFullRequest = queryTypeExecutor.getHeroGraphQLRequest(null);
 
 		// Of course, you can precise the fields you want
-		heroPartialRequest = queryType.getHeroGraphQLRequest("{appearsIn name}");
-		heroFriendsFriendsFriendsRequest = queryType
+		heroPartialRequest = queryTypeExecutor.getHeroGraphQLRequest("{appearsIn name}");
+		heroFriendsFriendsFriendsRequest = queryTypeExecutor
 				.getHeroGraphQLRequest("{id appearsIn friends {name friends {friends{id name appearsIn}}}}");
 
 		// The easiest way: don't precise which fields you want, and all known scalar fields are queried
-		humanFullRequest = queryType.getHumanGraphQLRequest(null);
+		humanFullRequest = queryTypeExecutor.getHumanGraphQLRequest(null);
 
 		// Of course, you can precise the fields you want
-		humanPartialRequest = queryType.getHumanGraphQLRequest("{appearsIn homePlanet name}");
-		humanFriendsFriendsFriendsRequest = queryType
+		humanPartialRequest = queryTypeExecutor.getHumanGraphQLRequest("{appearsIn homePlanet name}");
+		humanFriendsFriendsFriendsRequest = queryTypeExecutor
 				.getHumanGraphQLRequest("{id appearsIn name friends {name friends {friends{id name appearsIn}}}}");
 
 		// The easiest way: don't precise which fields you want, and all known scalar fields are queried
-		droidFullRequest = queryType.getDroidGraphQLRequest(null);
+		droidFullRequest = queryTypeExecutor.getDroidGraphQLRequest(null);
 
 		// Of course, you can precise the fields you want
-		droidPartialRequest = queryType.getDroidGraphQLRequest("{appearsIn primaryFunction name}");
-		droidFriendsFriendsFriendsRequest = queryType.getDroidGraphQLRequest(
+		droidPartialRequest = queryTypeExecutor.getDroidGraphQLRequest("{appearsIn primaryFunction name}");
+		droidFriendsFriendsFriendsRequest = queryTypeExecutor.getDroidGraphQLRequest(
 				"{id appearsIn name friends {name friends {friends{id name appearsIn}}} primaryFunction }");
 
 		// A demo of a wrong query. The preparation below should fail.
@@ -96,7 +91,7 @@ public class PartialPreparedRequests implements Queries {
 		// But here, we want the code to go on. So we just check to confirm that the exception occurs, and actually hide
 		// it.
 		try {
-			queryType.getDroidGraphQLRequest(
+			queryTypeExecutor.getDroidGraphQLRequest(
 					"{id appearsIn NON_EXISTING_FIELD friends {name friends {friends{id name appearsIn}}} primaryFunction }");
 
 			// Oups, if we got there, the expected exception was not thrown. Which means that the sample failed.
@@ -111,70 +106,70 @@ public class PartialPreparedRequests implements Queries {
 		}
 
 		// Mutations
-		createHumanRequest = mutationType
+		createHumanRequest = mutationTypeExecutor
 				.getCreateHumanGraphQLRequest("{id name homePlanet appearsIn friends {id name}}");
-		addFriendRequest = mutationType.getAddFriendGraphQLRequest("{id name appearsIn friends {id name}}");
+		addFriendRequest = mutationTypeExecutor.getAddFriendGraphQLRequest("{id name appearsIn friends {id name}}");
 	}
 
 	@Override
 	public Character heroFull() throws GraphQLRequestExecutionException {
-		return queryType.hero(heroFullRequest, null);
+		return queryTypeExecutor.hero(heroFullRequest, null);
 	}
 
 	@Override
 	public Character heroPartial(Episode episode) throws GraphQLRequestExecutionException {
-		return queryType.hero(heroPartialRequest, episode);
+		return queryTypeExecutor.hero(heroPartialRequest, episode);
 	}
 
 	@Override
 	public Character heroFriendsFriendsFriends(Episode episode) throws GraphQLRequestExecutionException {
-		return queryType.hero(heroFriendsFriendsFriendsRequest, episode);
+		return queryTypeExecutor.hero(heroFriendsFriendsFriendsRequest, episode);
 	}
 
 	@Override
 	public Human humanFull(String id) throws GraphQLRequestExecutionException {
-		return queryType.human(humanFullRequest, id);
+		return queryTypeExecutor.human(humanFullRequest, id);
 	}
 
 	@Override
 	public Human humanPartial(String id) throws GraphQLRequestExecutionException {
-		return queryType.human(humanPartialRequest, id);
+		return queryTypeExecutor.human(humanPartialRequest, id);
 	}
 
 	@Override
 	public Human humanFriendsFriendsFriends(String id) throws GraphQLRequestExecutionException {
-		return queryType.human(humanFriendsFriendsFriendsRequest, id);
+		return queryTypeExecutor.human(humanFriendsFriendsFriendsRequest, id);
 	}
 
 	@Override
 	public Droid droidFull(String id) throws GraphQLRequestExecutionException {
-		return queryType.droid(droidFullRequest, id);
+		return queryTypeExecutor.droid(droidFullRequest, id);
 	}
 
 	@Override
 	public Droid droidPartial(String id) throws GraphQLRequestExecutionException {
-		return queryType.droid(droidPartialRequest, id);
+		return queryTypeExecutor.droid(droidPartialRequest, id);
 	}
 
 	@Override
 	public Droid droidFriendsFriendsFriends(String id) throws GraphQLRequestExecutionException {
-		return queryType.droid(droidFriendsFriendsFriendsRequest, id);
+		return queryTypeExecutor.droid(droidFriendsFriendsFriendsRequest, id);
 	}
 
 	@Override
 	public Droid droidDoesNotExist() throws GraphQLRequestExecutionException {
-		return queryType.droid(droidFriendsFriendsFriendsRequest, "00000000-0000-0000-0000-000000001111");
+		return queryTypeExecutor.droid(droidFriendsFriendsFriendsRequest, "00000000-0000-0000-0000-000000001111");
 	}
 
 	@Override
 	public Human createHuman(String name, String homePlanet)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return mutationType.createHuman(createHumanRequest, name, homePlanet);
+		return mutationTypeExecutor.createHuman(createHumanRequest, name, homePlanet);
 	}
 
 	@Override
 	public Character addFriend(String idCharacter, String idNewFriend)
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
-		return mutationType.addFriend(addFriendRequest, idCharacter, idNewFriend);
+		return mutationTypeExecutor.addFriend(addFriendRequest, idCharacter, idNewFriend);
 	}
 }